+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.0] - 18.07.2018.
+
+- Methods in NTP are now static.
+- Added now() for returning current DateTime object with calculated NTP offset
+
+- getNtpTime was renamed to getNtpOffset (since it is returning int offset in milliseconds)
+
 ## [0.0.1] - 16.04.2018.
 
 - Get NTP clock offset